add base classes for task retriever
