+++ app/services/generator/media/image/retrieve_task/api_client.rb
@@ -2,25 +2,7 @@ module Generator
   module Media
     module Image
       module RetrieveTask
-        class ApiClient
-          def initialize(task_id, api_url)
-            @task_id = task_id
-            @api_url = api_url
-          end
-
-          def api_response
-            connection.get("#{api_url}/#{task_id}")
-          end
-
-          private
-
-          delegate :connection, to: :faraday_connection
-
-          def faraday_connection
-            ::Clients::Generator::Connection.new(api_url)
-          end
-
-          attr_reader :task_id, :api_url
+        class ApiClient < Generator::Media::RetrieveTask::ApiClientBase
         end
       end
     end
