+++ app/services/generator/media/image/retrieve_task/image_extractor.rb
@@ -2,26 +2,7 @@ module Generator
   module Media
     module Image
       module RetrieveTask
-        class ImageExtractor
-          include Memery
-
-          def initialize(response)
-            @response = response
-          end
-
-          def image_url
-            raise Freepik::ResponseError unless response.success?
-
-            parsed_body.dig("data", "generated", 0)
-          end
-
-          private
-
-          attr_reader :response
-
-          memoize def parsed_body
-            JSON.parse(response.body)
-          end
+        class ImageExtractor < Generator::Media::RetrieveTask::MediaExtractorBase
         end
       end
     end
