--- app/services/generator/media/image/retrieve_task/response_parser.rb
@@ -1,27 +1,0 @@
-module Generator
-  module Media
-    module Image
-      module RetrieveTask
-        class ResponseParser
-          include Memery
-
-          def initialize(raw_body)
-            @raw_body = raw_body
-          end
-
-          def image_url
-            parsed_body.dig("data", "generated", 0)
-          end
-
-          private
-
-          attr_reader :raw_body
-
-          memoize def parsed_body
-            JSON.parse(raw_body)
-          end
-        end
-      end
-    end
-  end
-end
