+++ app/services/generator/media/retrieve_task/api_client_base.rb
@@ -1,0 +1,26 @@
+module Generator
+  module Media
+    module RetrieveTask
+      class ApiClientBase
+        def initialize(task_id, api_url)
+          @task_id = task_id
+          @api_url = api_url
+        end
+
+        def api_response
+          connection.get("#{api_url}/#{task_id}")
+        end
+
+        private
+
+        delegate :connection, to: :faraday_connection
+
+        def faraday_connection
+          ::Clients::Generator::Connection.new(api_url)
+        end
+
+        attr_reader :task_id, :api_url
+      end
+    end
+  end
+end
