+++ app/services/generator/media/retrieve_task/api_url_fetcher_base.rb
@@ -1,0 +1,23 @@
+module Generator
+  module Media
+    module RetrieveTask
+      class ApiUrlFetcherBase
+        def initialize(processor)
+          @processor = processor
+        end
+
+        def api_url
+          api_urls[processor]
+        end
+
+        private
+
+        attr_reader :processor
+
+        def api_urls
+          self.class::API_URLS
+        end
+      end
+    end
+  end
+end
