+++ app/services/generator/media/retrieve_task/media_extractor_base.rb
@@ -1,0 +1,27 @@
+module Generator
+  module Media
+    module RetrieveTask
+      class MediaExtractorBase
+        include Memery
+
+        def initialize(response)
+          @response = response
+        end
+
+        def media_url
+          raise Freepik::ResponseError unless response.success?
+
+          parsed_body.dig("data", "generated", 0)
+        end
+
+        private
+
+        attr_reader :response
+
+        memoize def parsed_body
+          JSON.parse(response.body)
+        end
+      end
+    end
+  end
+end
